+++ test/asset.test.ts
@@ -1,8 +1,8 @@
 import { generateAssetName, generateAssetSymbol } from "./util/util";
-import { Connection } from "../client/lib/ft4/types";
+import { Connection } from "/ft4/types";
 import { createChromiaClient, getNewAsset } from "./util/blockchain-util";
-import { InvalidUrlError } from "../client/lib/ft4/asset/interfaces";
-import { createConnection } from "../client/lib/ft4/ft-session";
+import { InvalidUrlError } from "/ft4/asset/interfaces";
+import { createConnection } from "/ft4/ft-session";
 import { Buffer } from "buffer";
 import { IClient, gtv } from "postchain-client";
 import { randomBytes } from "crypto";