+++ test/key-store-interactor.test.ts
@@ -1,11 +1,11 @@
 import { newSignatureProvider } from "postchain-client";
-import { FlagsType } from "../client/lib/ft4/accounts/auth-descriptor";
-import { createInMemoryFtKeyStore } from "../client/lib/ft4/authentication/ft/key-stores/in-memory";
+import { FlagsType } from "/ft4/accounts/auth-descriptor";
+import { createInMemoryFtKeyStore } from "/ft4/authentication/ft/key-stores/in-memory";
 import {
   KeyStoreInteractor,
   createConnection,
   createKeyStoreInteractor,
-} from "../client/lib/ft4/ft-session";
+} from "/ft4/ft-session";
 import AccountBuilder from "./util/account-builder";
 import { Connection } from "/ft4/types";
 import { createChromiaClient } from "./util/blockchain-util";